feat(invoice-details): add routes and restructure file location
